+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.8.0] - 2023-06-15
+
 ### Added
 
 - Support for featureType in JSON to GeoJSON and GeoJSON to JSON